large docs refactoring (#172)
* large docs refactoring

* large docs refactoring

* large docs refactoring

* test fixes

* test fixes + changelog

* test fixes + minidoc fixes

* test fixes + minidoc fixes

* LISTA fixes

* LISTA fixes

* fixed issue in the definition of ScorePrior

* improved Blur performance and added warning to tomography about normalization

* fix blur

* fix blur

* DPIR params fix

* DPIR params fix

* DPIR params fix

* add plot interpolation (nice for showing missing pixels in inpainted images)

* add plot interpolation (nice for showing missing pixels in inpainted images)

* fixed preconULA

* wavelet prior

* fix doc

---------
