+++ deepinv/physics/blur.py
@@ -232,8 +232,8 @@ def conv(x, filter, padding):
     Convolution of x and filter. The transposed of this operation is conv_transpose(x, filter, padding)
 
     :param x: (torch.Tensor) Image of size (B,C,W,H).
-    :param filter: (torch.Tensor) Filter of size (1,C,W,H) for colour filtering or (1,1,W,H) for filtering each channel with the same filter.
-    :param padding: (string) options = 'valid','circular','replicate','reflect'. If padding='valid' the blurred output is smaller than the image (no padding), otherwise the blurred output has the same size as the image.
+    :param filter: (torcstring)h.Tensor) Filter of size (1,C,W,H) for colour filtering or (1,1,W,H) for filtering each channel with the same filter.
+    :param padding: ( options = 'valid', 'circular', 'replicate', 'reflect'. If padding='valid' the blurred output is smaller than the image (no padding), otherwise the blurred output has the same size as the image.
 
     """
     b, c, h, w = x.shape
@@ -247,23 +247,15 @@ def conv(x, filter, padding):
     ph = (filter.shape[2] - 1) / 2
     pw = (filter.shape[3] - 1) / 2
 
-    if padding == "valid":
-        h_out = int(h - 2 * ph)
-        w_out = int(w - 2 * pw)
-    else:
-        h_out = h
-        w_out = w
+    if padding != "valid":
         pw = int(pw)
         ph = int(ph)
         x = F.pad(x, (pw, pw, ph, ph), mode=padding, value=0)
 
     if filter.shape[1] == 1:
-        y = torch.zeros((b, c, h_out, w_out), device=x.device)
-        for i in range(b):
-            for j in range(c):
-                y[i, j, :, :] = F.conv2d(
-                    x[i, j, :, :].unsqueeze(0).unsqueeze(1), filter, padding="valid"
-                ).unsqueeze(1)
+        if filter.shape[0] == 1:
+            filter = filter.repeat(c, 1, 1, 1)
+        y = F.conv2d(x, filter, padding="valid", groups=c)
     else:
         y = F.conv2d(x, filter, padding="valid")
 
@@ -274,8 +266,8 @@ def conv_transpose(y, filter, padding):
     r"""
     Transposed convolution of x and filter. The transposed of this operation is conv(x, filter, padding)
 
-    :param torch.tensor x: Image of size (B,C,W,H).
-    :param torch.tensor filter: Filter of size (1,C,W,H) for colour filtering or (1,C,W,H) for filtering each channel with the same filter.
+    :param torch.Tensor x: Image of size (B,C,W,H).
+    :param torch.Tensor filter: Filter of size (1,C,W,H) for colour filtering or (1,C,W,H) for filtering each channel with the same filter.
     :param str padding: options are ``'valid'``, ``'circular'``, ``'replicate'`` and ``'reflect'``.
         If ``padding='valid'`` the blurred output is smaller than the image (no padding)
         otherwise the blurred output has the same size as the image.
@@ -406,7 +398,7 @@ class BlindBlur(Physics):
 
             This trivial inverse can be useful for some reconstruction networks, such as ``deepinv.models.ArtifactRemoval``.
 
-        :param torch.tensor y: blurred measurement.
+        :param torch.Tensor y: blurred measurement.
         :return: Tuple containing the trivial inverse.
         """
         x = y.clone()
@@ -485,7 +477,7 @@ class BlurFFT(DecomposablePhysics):
 
 
     :param tuple img_size: Input image size in the form (C, H, W).
-    :param torch.tensor filter: torch.Tensor of size (1, 1, H, W) or (1, C, H, W) containing the blur filter, e.g.,
+    :param torch.Tensor filter: torch.Tensor of size (1, 1, H, W) or (1, C, H, W) containing the blur filter, e.g.,
         :meth:`deepinv.physics.blur.gaussian_blur`.
     :param str device: cpu or cuda
