+++ deepinv/tests/test_loss_train.py
@@ -204,7 +204,7 @@ def test_train_patchnr(imsize, dummy_dataset, device):
     patchnr_dataloader = DataLoader(
         patch_dataset, batch_size=32, shuffle=True, drop_last=True
     )
-    patchnr = dinv.models.PatchNR(
+    patchnr = dinv.optim.PatchNR(
         channels=test_sample.shape[1],
         patch_size=patch_size,
         sub_net_size=64,
@@ -228,7 +228,7 @@ def test_epll_parameter_estimation(imsize, dummy_dataset, device):
     patch_dataloader = torch.utils.data.DataLoader(
         patch_dataset, batch_size=2, shuffle=True, drop_last=False
     )
-    epll = dinv.models.EPLL(channels=imsize[0], pretrained=None, n_components=3)
+    epll = dinv.optim.EPLL(channels=imsize[0], pretrained=None, n_components=3)
     epll.GMM.fit(patch_dataloader, max_iters=10)
 
     assert not torch.any(torch.isnan(epll.GMM.mu))
