+++ deepinv/tests/test_models.py
@@ -79,7 +79,7 @@ def choose_denoiser(name, imsize):
     elif name == "swinir":
         out = dinv.models.SwinIR(in_chans=imsize[0])
     elif name == "epll":
-        out = dinv.models.EPLL(channels=imsize[0])
+        out = dinv.models.EPLLDenoiser(channels=imsize[0])
     else:
         raise Exception("Unknown denoiser")
 
@@ -184,7 +184,7 @@ def test_wavelet_models_identity():
         "git+https://github.com/fbcotter/pytorch_wavelets.git`",
     )
 
-    # 1. Wavelet Prior & dictionary
+    # 1. Wavelet denoiser (single & dictionary)
     for dimension in ["2d", "3d"]:
         wvdim = 2 if dimension == "2d" else 3
         x = (
@@ -211,6 +211,24 @@ def test_wavelet_models_identity():
         x_hat = model(x, 0.0)
         assert x_hat.shape == x.shape
         assert torch.allclose(x, x_hat, atol=1e-5)  # The model should be the identity
+
+    # 2. Wavelet Prior
+    for dimension in ["2d", "3d"]:
+        wvdim = 2 if dimension == "2d" else 3
+        x = (
+            torch.ones((4, 3, 31, 27))
+            if dimension == "2d"
+            else torch.ones((4, 3, 31, 27, 29))
+        )
+        level = 3
+        prior = dinv.optim.prior.WaveletPrior(wvdim=wvdim, p=1, level=level)
+        g_nonflat = prior.g(x, reduce=False)
+        g_flat = prior.g(x, reduce=True)
+        assert g_nonflat.dim() > 0
+        assert len(g_nonflat) == 3 * level if wvdim == 2 else 7 * level
+        assert g_flat.dim() == 0
+
+        assert torch.allclose(g_nonflat.abs().sum(), g_flat)
 
 
 def test_TV_models_identity():
