+++ docs/source/deepinv.denoisers.rst
@@ -1,33 +1,30 @@
-.. _models:
+.. _denoisers:
 
-Models
-======
-This package provides vanilla signal reconstruction methods,
-which can be used for a quick evaluation of a learning setting.
-
-
-.. autosummary::
-   :toctree: stubs
-   :template: myclass_template.rst
-   :nosignatures:
-
-   deepinv.models.ArtifactRemoval
-   deepinv.models.DeepImagePrior
 
 Denoisers
----------
+==================
+
 Denoisers are :class:`torch.nn.Module` that take a noisy image as input and return a denoised image.
 They can be used as a building block for plug-and-play restoration, for building unrolled architectures,
 or as a standalone denoiser. All denoisers have a ``forward`` method that takes a noisy image and a noise level
-(which generally corresponds to the standard deviation of the noise) as input and returns a denoised image.
+(which generally corresponds to the standard deviation of the noise) as input and returns a denoised image:
+
+    >>> import torch
+    >>> import deepinv as dinv
+    >>> denoiser = dinv.models.DRUNet()
+    >>> sigma = 0.1
+    >>> image = torch.ones(1, 3, 32, 32)*.5
+    >>> noisy_image =  image + torch.randn(1, 3, 32, 32) * sigma
+    >>> denoised_image = denoiser(noisy_image, sigma)
 
 .. note::
 
     Some denoisers (e.g., :class:`deepinv.models.DnCNN`) do not use the information about the noise level.
     In this case, the noise level is ignored.
 
+
 Classical Denoisers
-^^^^^^^^^^^^^^^^^^^^^
+-------------------
 
 .. autosummary::
    :toctree: stubs
@@ -40,10 +37,11 @@ Classical Denoisers
    deepinv.models.TGVDenoiser
    deepinv.models.WaveletDenoiser
    deepinv.models.WaveletDictDenoiser
+   deepinv.models.EPLLDenoiser
 
 
-Learnable Denoisers
-^^^^^^^^^^^^^^^^^^^^^
+Deep Denoisers
+-------------------
 
 .. autosummary::
    :toctree: stubs
@@ -51,7 +49,6 @@ Learnable Denoisers
    :nosignatures:
 
    deepinv.models.AutoEncoder
-   deepinv.models.ConvDecoder
    deepinv.models.UNet
    deepinv.models.DnCNN
    deepinv.models.DRUNet
@@ -61,8 +58,9 @@ Learnable Denoisers
    deepinv.models.DiffUNet
 
 
-Equivariant denoisers
----------------------
+
+Equivariant Denoisers
+--------------------------
 The denoisers can be turned into equivariant denoisers by wrapping them with the
 :class:`deepinv.models.EquivariantDenoiser` class.
 The group of transformations available at the moment are vertical/horizontal flips, 90 degree rotations, or a
@@ -80,42 +78,9 @@ equivariant denoiser.
    deepinv.models.EquivariantDenoiser
 
 
-Unfolded architectures
-----------------------
-Some more specific unfolded architectures are also available.
-
-.. autosummary::
-   :toctree: stubs
-   :template: myclass_template.rst
-   :nosignatures:
-
-   deepinv.models.PDNet_PrimalBlock
-   deepinv.models.PDNet_DualBlock
-
-Patch Priors
-------------
-Explicit priors using statistical models on the space of patches.
-
-.. autosummary::
-   :toctree: stubs
-   :template: myclass_template.rst
-   :nosignatures:
-
-    deepinv.models.EPLL
-    deepinv.models.PatchNR
-
-EPLL relys on Gaussian mixture models (GMMs). The following class contains an expectation maximization algorithm for
-parameter estimation in GMMs.
-
-.. autosummary::
-   :toctree: stubs
-   :template: myclass_template.rst
-   :nosignatures: 
-    
-    deepinv.models.epll.GaussianMixtureModel
 
 .. _pretrained-weights:
-Pretrained weights
+Pretrained Weights
 ------------------
 The following denoisers have **pretrained weights** available; we next briefly summarize the origin of the weights,
 associated reference and relevant details. All pretrained weights are hosted on
@@ -159,3 +124,5 @@ associated reference and relevant details. All pretrained weights are hosted on
      - Default: parameters estimated with deepinv on 50 mio patches from the training/validation images from BSDS500 for grayscale and color images.
    * - 
      - Code for generating the weights for the example :ref:`patch-prior-demo` is contained within the demo
+
+
