+++ examples/patch-priors/demo_epll.py
@@ -5,10 +5,11 @@ Expected Patch Log Likelihood (EPLL) for Denoising and Inpainting
 In this example we use the expected patch log likelihood (EPLL) prior EPLL proposed in `"From learning models of natural image patches to whole image restoration" <https://ieeexplore.ieee.org/document/6126278>`_. 
 for denoising and inpainting of natural images.
 To this end, we consider the inverse problem :math:`y = Ax+\epsilon`, where :math:`A` is either the identity (for denoising)
-or a masking operator (for inpainting) and :math:`\epsilon\sim\mathcal{N}(0,\sigma^2 I)` is white Gaussian noise with standard deviation :math`\sigma`.
+or a masking operator (for inpainting) and :math:`\epsilon\sim\mathcal{N}(0,\sigma^2 I)` is white Gaussian noise with standard deviation :math:`\sigma`.
 """
 
-from deepinv.models import EPLL
+from deepinv.models import EPLLDenoiser
+from deepinv.optim import EPLL
 from deepinv.physics import GaussianNoise, Denoising, Inpainting
 from deepinv.utils import cal_psnr, plot
 import torch
@@ -26,7 +27,7 @@ device = "cuda" if torch.cuda.is_available() else "cpu"
 url = get_image_url("CBSD_0010.png")
 test_img = load_url_image(url, grayscale=False).to(device)
 patch_size = 6
-model_EPLL = EPLL(channels=test_img.shape[1], patch_size=patch_size, device=device)
+model = EPLL(channels=test_img.shape[1], patch_size=patch_size, device=device)
 
 # %%
 # Denoising
@@ -34,7 +35,7 @@ model_EPLL = EPLL(channels=test_img.shape[1], patch_size=patch_size, device=devi
 # In this setting, the operator :math:`A` is the identity; we set the noise level to :math:`\sigma=25/255`.
 # Define noise model, operator and generate observation
 
-sigma = 25.0 / 255.0
+sigma = 0.1
 noise_model = GaussianNoise(sigma)
 physics = Denoising(device=device, noise_model=noise_model)
 observation = physics(test_img)
@@ -42,11 +43,11 @@ observation = physics(test_img)
 # %%
 # We use the default choice of the betas in the half quadratic splitting given by
 # :math:`\beta \in \sigma^{-2} \{1,4,8,16,32\}`.
-# Generally, the betas are hyperparameters, which have to be chosen for each inverse problem seperately.
+# Generally, the betas are hyperparameters, which have to be chosen for each inverse problem separately.
 #
 
 # Reconstruction
-x_out = model_EPLL(observation, sigma, batch_size=5000)
+x_out = model(observation, physics, batch_size=5000)
 
 # PSNR computation and plots.
 psnr_obs = cal_psnr(observation, test_img)
@@ -68,8 +69,9 @@ plot(
 # Define noise model, operator and generate observation
 
 sigma = 0.01
-noise_model = GaussianNoise(sigma)
-physics = Inpainting(test_img[0].shape, mask=0.7, device=device)
+physics = Inpainting(
+    test_img[0].shape, mask=0.7, device=device, noise_model=GaussianNoise(sigma)
+)
 observation = physics(test_img)
 
 # %%
@@ -79,9 +81,7 @@ observation = physics(test_img)
 betas = [1.0, 5.0, 10.0, 40.0, 80.0, 160.0, 320.0]
 
 # Reconstruction
-x_out = model_EPLL.reconstruction(
-    observation, observation.clone(), sigma, physics, betas=betas, batch_size=5000
-)
+x_out = model(observation, physics, betas=betas, batch_size=5000)
 
 # PSNR computation and plots
 psnr_obs = cal_psnr(observation, test_img)
